feat(database): Migration to Postgres
- Switched to a more solid database - Postgres;
- Created appsettings example to make repo cloning easier;
- Created Initial Migration and DB seeding for the Postgres database;
- Added db seeding for Roles, Admin User and Address Types;
- Changed the AutoIncrement to UseIdentityByDefaultColumn;

+++ Data/ModelConfiguration/ProductAvaliationModelConfiguration.cs
@@ -13,10 +13,10 @@ public class ProductAvaliationModelConfiguration : IEntityTypeConfiguration<Prod
         builder.Property(x => x.Description)
             .IsRequired()
             .HasMaxLength(500);
-        
+
         builder.Property(x => x.Stars)
             .IsRequired()
-            .HasColumnType("TINYINT");
+            .HasColumnType("smallint");
 
         builder.HasOne(x => x.Client)
             .WithMany(y => y.ProductAvaliations)
